#1458 - Move level field to global #1461 - Knight is no longer affected by knockback sources while using bulls charge
